Moved TempStyle, PendingDeprecation to utils.py
